refactor: 이벤트 핸들러 경로 변경
- application/service -> application/event
