+++ src/main/java/followMe/hub_server/stock/application/event/OrderEventHandler.java
@@ -1,10 +1,9 @@
-package followMe.hub_server.stock.application.service;
+package followMe.hub_server.stock.application.event;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.followMe.common.event.inbox.Inbox;
 import com.followMe.common.event.inbox.InboxRepository;
-import followMe.hub_server.stock.application.event.OrderCancelledEvent;
-import followMe.hub_server.stock.application.event.OrderRequestedEvent;
+import followMe.hub_server.stock.application.service.HubStockService;
 import java.util.Map;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
